Permit overwrite of MdxNode field addition to page context
The block of code doesn't work in conjunction with some customization (e.g. when `page.component` is defined to point to a template or when pages are not in `src/pages`. Adding a blog to avoid re-running the block of code if the MdxNode has been already defined will permit to create additional `onCreatePage` method to apply custom logic.

+++ packages/gatsby-theme-carbon/gatsby-node.js
@@ -25,24 +25,26 @@ exports.onCreatePage = (
   { page, actions, getNodesByType, ...rest },
   pluginOptions
 ) => {
-  // Find the MdxNode that created our page
-  const MdxNode = getNodesByType('Mdx').find(
-    ({ fileAbsolutePath }) => fileAbsolutePath === page.component
-  );
+  if(!page.context.MdxNode){ // Enable the possibility to overwrite it
+    // Find the MdxNode that created our page
+    const MdxNode = getNodesByType('Mdx').find(
+      ({ fileAbsolutePath }) => fileAbsolutePath === page.component
+    );
 
-  const { titleType = 'page' } = pluginOptions;
-  const { createPage, deletePage } = actions;
-  const [relativePagePath] = page.componentPath.split('src/pages').splice('-1');
-  deletePage(page);
-  createPage({
-    ...page,
-    context: {
-      ...page.context,
-      relativePagePath,
-      titleType,
-      MdxNode,
-    },
-  });
+    const { titleType = 'page' } = pluginOptions;
+    const { createPage, deletePage } = actions;
+    const [relativePagePath] = page.componentPath.split('src/pages').splice('-1');
+    deletePage(page);
+    createPage({
+      ...page,
+      context: {
+        ...page.context,
+        relativePagePath,
+        titleType,
+        MdxNode,
+      },
+    });
+  }
 };
 
 // In the case where none of the nav items have child pages, an error occurs in the left nav
